# optimize negate