--- a/int128_test.go
+++ b/int128_test.go
@@ -95,6 +95,9 @@ func FuzzInt128_String(f *testing.F) {
 	f.Add(uint64(0), uint64(0))
 	f.Add(uint64(1), uint64(0))
 	f.Add(uint64(math.MaxUint64), uint64(math.MaxUint64))
+	f.Add(uint64(1<<63), uint64(0))
+	f.Add(uint64(1<<63), uint64(1))
+
 	f.Fuzz(func(t *testing.T, u0, u1 uint64) {
 		a := Int128{u0, u1}
 		got := a.String()