--- a/int256_test.go
+++ b/int256_test.go
@@ -123,6 +123,7 @@ func FuzzInt256_String(f *testing.F) {
 	f.Add(uint64(0), uint64(1), uint64(0), uint64(0))
 	f.Add(uint64(1), uint64(0), uint64(0), uint64(0))
 	f.Add(uint64(math.MaxUint64), uint64(math.MaxUint64), uint64(math.MaxUint64), uint64(math.MaxUint64))
+	f.Add(uint64(1<<63), uint64(0), uint64(0), uint64(0))
 	f.Fuzz(func(t *testing.T, u0, u1, u2, u3 uint64) {
 		a := Int256{u0, u1, u2, u3}
 		got := a.String()