```plain
% benchstat .old.txt .new.txt                                                                                                                            fix-benchmark-of-text
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/ints
cpu: Apple M1 Pro
                   │  .old.txt   │              .new.txt              │
                   │   sec/op    │   sec/op     vs base               │
Int1024_Text2-10     11.57µ ± 0%   11.57µ ± 0%       ~ (p=0.644 n=10)
Int1024_Text10-10    12.18µ ± 0%   12.18µ ± 0%       ~ (p=0.698 n=10)
Int1024_Text62-10    11.57µ ± 0%   11.56µ ± 0%  -0.09% (p=0.011 n=10)
Int1024_String-10    12.38µ ± 1%   12.20µ ± 1%  -1.51% (p=0.001 n=10)
Int128_Text2-10      429.1n ± 2%   422.0n ± 0%  -1.65% (p=0.000 n=10)
Int128_Text10-10     217.8n ± 2%   214.8n ± 0%  -1.38% (p=0.000 n=10)
Int128_Text62-10     121.7n ± 1%   120.4n ± 0%  -1.07% (p=0.000 n=10)
Int128_String-10     220.5n ± 2%   213.7n ± 0%  -3.08% (p=0.000 n=10)
Int256_Text2-10      1.048µ ± 1%   1.031µ ± 0%  -1.58% (p=0.000 n=10)
Int256_Text10-10     700.8n ± 3%   692.2n ± 0%  -1.23% (p=0.000 n=10)
Int256_Text62-10     406.2n ± 2%   401.5n ± 0%  -1.17% (p=0.000 n=10)
Int256_String-10     699.3n ± 1%   691.7n ± 0%  -1.09% (p=0.000 n=10)
Int512_Text2-10      3.124µ ± 3%   3.082µ ± 1%  -1.33% (p=0.002 n=10)
Int512_Text10-10     2.584µ ± 3%   2.547µ ± 1%  -1.45% (p=0.000 n=10)
Int512_Text62-10     1.680µ ± 2%   1.664µ ± 1%  -0.95% (p=0.013 n=10)
Int512_String-10     2.579µ ± 1%   2.542µ ± 0%  -1.45% (p=0.000 n=10)
Uint1024_Text2-10    43.60µ ± 2%   43.26µ ± 0%  -0.78% (p=0.000 n=10)
Uint1024_Text10-10   12.54µ ± 2%   12.36µ ± 0%  -1.43% (p=0.000 n=10)
Uint1024_Text62-10   8.449µ ± 2%   8.299µ ± 0%  -1.78% (p=0.000 n=10)
Uint1024_String-10   12.57µ ± 2%   12.36µ ± 0%  -1.72% (p=0.000 n=10)
Uint128_Text2-10     645.9n ± 3%   643.5n ± 0%  -0.38% (p=0.002 n=10)
Uint128_Text10-10    223.2n ± 3%   218.5n ± 0%  -2.13% (p=0.000 n=10)
Uint128_Text62-10    128.1n ± 4%   125.2n ± 0%  -2.23% (p=0.000 n=10)
Uint128_String-10    218.2n ± 1%   217.6n ± 0%       ~ (p=0.092 n=10)
Uint256_Text2-10     2.061µ ± 0%   2.059µ ± 0%       ~ (p=0.467 n=10)
Uint256_Text10-10    677.3n ± 0%   677.0n ± 0%       ~ (p=0.362 n=10)
Uint256_Text62-10    402.4n ± 0%   402.3n ± 0%       ~ (p=0.253 n=10)
Uint256_String-10    677.2n ± 0%   676.8n ± 0%       ~ (p=0.092 n=10)
Uint512_Text2-10     8.506µ ± 0%   8.504µ ± 0%       ~ (p=0.146 n=10)
Uint512_Text10-10    2.659µ ± 0%   2.685µ ± 1%       ~ (p=0.092 n=10)
Uint512_Text62-10    1.645µ ± 0%   1.664µ ± 3%  +1.19% (p=0.000 n=10)
Uint512_String-10    2.668µ ± 0%   2.702µ ± 1%  +1.27% (p=0.000 n=10)
geomean              1.598µ        1.585µ       -0.83%
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Refactor**
  - Improved internal handling of negative multi-word unsigned integers for large number formatting (128-bit, 256-bit, 512-bit, and 1024-bit values). No changes to visible features or exported interfaces.
- **Tests**
  - Expanded fuzz testing with new edge case inputs for 128-bit and 256-bit integer string conversions to enhance reliability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->